+++ calingen/contrib/holidays/germany/provider.py
@@ -15,8 +15,8 @@ from dateutil.rrule import WE, YEARLY, rrule
 from calingen.constants import EventCategory
 from calingen.interfaces.data_exchange import (
     SOURCE_EXTERNAL,
-    CalenderEntry,
-    CalenderEntryList,
+    CalendarEntry,
+    CalendarEntryList,
 )
 from calingen.interfaces.plugin_api import EventProvider
 
@@ -155,10 +155,10 @@ class GermanyFederal(EventProvider):
     @classmethod
     def resolve(cls, year):  # noqa: D102
 
-        result = CalenderEntryList()
+        result = CalendarEntryList()
         for i in cls.holidays:
             result.add(
-                CalenderEntry(
+                CalendarEntry(
                     i[0],
                     i[1],
                     i[2].between(