+++ calingen/views/web.py
@@ -6,7 +6,7 @@ These views are not model-specific (see
 :mod:`calingen.views.event` and :mod:`calingen.views.profile`), but provide
 representations of the app's data.
 
-Please note: The focus of calingen is to create analogous, paper-based calender
+Please note: The focus of calingen is to create analogous, paper-based calendar
 pages. These views are provided as convenience!
 """
 
@@ -15,13 +15,13 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.views.generic.base import TemplateView
 
 # app imports
-from calingen.views.mixins import AllCalenderEntriesMixin, RestrictToUserMixin
+from calingen.views.mixins import AllCalendarEntriesMixin, RestrictToUserMixin
 
 
-class CalenderEntryListView(
+class CalendarEntryListView(
     LoginRequiredMixin,
     RestrictToUserMixin,
-    AllCalenderEntriesMixin,
+    AllCalendarEntriesMixin,
     TemplateView,
 ):
     """Provide a list view of all events in a given year.
@@ -34,4 +34,4 @@ class CalenderEntryListView(
     optimization.
     """
 
-    template_name = "calingen/calender_entry_list_year.html"
+    template_name = "calingen/calendar_entry_list_year.html"